+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- `firebase emulator:start` use a default project if no project can be found. (#9072)
+- `firebase emulator:start` use a default project `demo-no-project` if no project can be found. (#9072)
+- `firebase init dataconnect` also supports bootstrapping flutter template. (#9084)